Registration form update. Automatic login after registration
