+++ src/pages/register/RegistrationForm.css
@@ -1,0 +1,12 @@
+.formContainer {
+    display: flex;
+    width: 100%;
+    justify-content: center;
+    align-items: center;
+}
+
+.formContainer>form {
+    width: 80%;
+    max-width: 600px;
+    margin-top: 40px;
+}
